+++ src/components/index.js
@@ -109,6 +109,9 @@ export ContentsTagsModal from '@plone/volto/components/manage/Contents/ContentsT
 export UsersControlpanelUser from '@plone/volto/components/manage/Controlpanels/UsersControlpanelUser';
 export UsersControlpanelGroups from '@plone/volto/components/manage/Controlpanels/UsersControlpanelGroups';
 export DiffField from '@plone/volto/components/manage/Diff/DiffField';
+export DragDropList from '@plone/volto/components/manage/DragDropList/DragDropList';
+export InlineForm from '@plone/volto/components/manage/Form/InlineForm';
+export BlocksForm from '@plone/volto/components/manage/Blocks/Block/BlocksForm';
 
 export FormFieldWrapper from '@plone/volto/components/manage/Widgets/FormFieldWrapper';
 export ArrayWidget from '@plone/volto/components/manage/Widgets/ArrayWidget';